requirements.txt
alibabacloud_alidns20150109==3.5.6
alibabacloud_tea_openapi==0.3.12
alibabacloud_tea_util==0.3.13
loguru==0.7.3
PyYAML==6.0.2
Requests==2.32.3
